# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
        "api/artur/barberShop": {
            "post": {
                "tags": [
                    "Artur"
                ],
                "summary": "Cria um novo Barber Shop",
                "operationId": "createBarberShop",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "descripition",
                                    "advantages",
                                    "contact"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Barber Shop A"
                                    },
                                    "descripition": {
                                        "type": "string",
                                        "example": "Melhor Barber Shop da cidade"
                                    },
                                    "advantages": {
                                        "type": "string",
                                        "example": "Preços acessíveis, ambiente agradável"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": "1234-5678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barber Shop criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "barberShop": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Barber Shop A"
                                                },
                                                "descripition": {
                                                    "type": "string",
                                                    "example": "Melhor Barber Shop da cidade"
                                                },
                                                "advantages": {
                                                    "type": "string",
                                                    "example": "Preços acessíveis, ambiente agradável"
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "example": "1234-5678"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Barber Shop criado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/artur/barberShop/{id}": {
            "get": {
                "tags": [
                    "Artur"
                ],
                "summary": "Exibe as informações de um Barber Shop específico",
                "operationId": "getBarberShop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Barber Shop",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do Barber Shop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "barberShop": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Barber Shop A"
                                                },
                                                "descripition": {
                                                    "type": "string",
                                                    "example": "Melhor Barber Shop da cidade"
                                                },
                                                "advantages": {
                                                    "type": "string",
                                                    "example": "Preços acessíveis, ambiente agradável"
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "example": "1234-5678"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Produto A"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 29.99
                                                    },
                                                    "barber_shop_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-21T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-21T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Artur"
                ],
                "summary": "Atualiza as informações de um Barber Shop existente",
                "operationId": "updateBarberShop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Barber Shop",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "descripition",
                                    "advantages",
                                    "contact"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Barber Shop Atualizado"
                                    },
                                    "descripition": {
                                        "type": "string",
                                        "example": "Descrição atualizada"
                                    },
                                    "advantages": {
                                        "type": "string",
                                        "example": "Novas vantagens"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": "9876-5432"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barber Shop atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "barberShop": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Barber Shop Atualizado"
                                                },
                                                "descripition": {
                                                    "type": "string",
                                                    "example": "Descrição atualizada"
                                                },
                                                "advantages": {
                                                    "type": "string",
                                                    "example": "Novas vantagens"
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "example": "9876-5432"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-21T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Barber Shop atualizado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artur"
                ],
                "summary": "Remove um Barber Shop existente",
                "operationId": "deleteBarberShop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Barber Shop",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barber Shop deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Barber Shop deletado com sucesso"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}